Fixes two compatibility issues that makes bowser play nicer with others
When the 'define' function exists, bowser tried to register itself
against it. The problem is that it does not supply a name for the
definitation, causing it to be lost forever.

The second problem has to do with semi-colons. When the source file
is minified with uglify it breaks horibly. I've added ending semi-
colons where to ending control statements, which allows it to be
compatible with the majority of other compressors.