Fix Chrome detection for Android 8 OPR6 build num
One of Android 8's build numbers is 'opr6.170623.013'. This is expressed
in the user agent as:

"Mozilla/5.0 (Linux; Android 8.0.0; Nexus 6P Build/OPR6.170623.013)
AppleWebKit/537.36 (KHTML, like Gecko) Chrome/60.0.3112.107 Mobile
Safari/537.36"

Because this library detects the Opera browser if "opera", "opr", or
"opios" is contained in the user agent, the above user agent string is
incorrectly detected as a versionless Opera, instead of Chrome 60.

This commit changes the detection of "opr" to require a trailing slash
'/' character. This works because uild numbers are expressed as
Build/X..., whereas Opera versions are expressed as OPR/X.